Third pass at refactoring into ASP.NET Core
This time everything is tested and works, and a nuget package is being generated.
