+++ Toolhouse.Monitoring.Tests/packages.config
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net461" />
+  <package id="NUnit" version="3.6.0" targetFramework="net452" />
+  <package id="prometheus-net" version="1.2.4.0" targetFramework="net461" />
+  <package id="protobuf-net" version="2.0.0.668" targetFramework="net452" />
+  <package id="Rx-Core" version="2.2.5" targetFramework="net461" />
+  <package id="Rx-Interfaces" version="2.2.5" targetFramework="net461" />
+</packages>
