--- src/Toolhouse.Monitoring.Core/Dependencies/IDependency.cs
@@ -1,19 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Toolhouse.Monitoring.Core.Dependencies
-{
-    public interface IDependency
-    {
-        /// <summary>
-        /// Short name of this dependency.
-        /// </summary>
-        string Name { get; }
-
-        /// <summary />
-        /// <returns></returns>
-        DependencyStatus Check();
-    }
-}
