--- src/Toolhouse.Monitoring.Net461/Properties/AssemblyInfo.cs
@@ -1,36 +1,0 @@
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Toolhouse.Monitoring")]
-[assembly: AssemblyDescription("Monitoring tools for ASP.NET websites.")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Toolhouse")]
-[assembly: AssemblyProduct("Toolhouse.Monitoring.Properties")]
-[assembly: AssemblyCopyright("2017")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("d416b3d4-6f91-4c92-af08-4dcbedecb75a")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version
-//      Build Number
-//      Revision
-//
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
-
-// Nuget prefers AssemblyInformationalVersion, so modify this when cutting Nuget releases
-[assembly: AssemblyInformationalVersion("1.2.0")]
