--- src/Toolhouse.Monitoring.Net461/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="6.0.3" targetFramework="net461" />
-  <package id="prometheus-net" version="1.3.6-rc" targetFramework="net461" />
-  <package id="protobuf-net" version="2.1.0" targetFramework="net461" />
-  <package id="Rx-Core" version="2.2.5" targetFramework="net461" />
-  <package id="Rx-Interfaces" version="2.2.5" targetFramework="net461" />
-  <package id="System.Reactive" version="4.0.0-preview00001" targetFramework="net461" />
-</packages>
