--- src/Toolhouse.Monitoring.NetCore/HttpAuthChecker.cs
@@ -1,26 +1,0 @@
-using Microsoft.AspNetCore.Http;
-using Toolhouse.Monitoring.Core;
-
-namespace Toolhouse.Monitoring.NetCore
-{
-    public class HttpAuthChecker : AuthChecker
-    {
-        public HttpAuthChecker(HttpContext context, string username, string passwordSha256)
-            : base(username, passwordSha256)
-        {
-            _context = context;
-        }
-
-        protected override string GetAuthHeader()
-        {
-            return _context.Request.Headers["Authorization"];
-        }
-
-        protected override void OnInvalidAuthHeader()
-        {
-            _context.Response.StatusCode = 403;
-        }
-
-        HttpContext _context;
-    }
-}
