--- src/Toolhouse.Monitoring.NetCore/Toolhouse.Monitoring.NetCore.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.1-beta1" />
-    <PackageReference Include="prometheus-net" Version="1.3.6-rc" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Toolhouse.Monitoring.Core\Toolhouse.Monitoring.Core.csproj" />
-  </ItemGroup>
-
-</Project>
